Ads c# to updates series data vs updating the whole chart definition. And limit table & json logging to first pass only

+++ tapeStream/Client/Components/HighCharts/StackedColumns3D.model.cs
@@ -79,8 +79,8 @@ namespace tapeStream.Client.Components.HighCharts
     public class Plotband
     {
         public string color { get; set; }
-        public int from { get; set; }
-        public int to { get; set; }
+        public decimal from { get; set; }
+        public decimal to { get; set; }
         public Label label { get; set; }
     }
